Add timepoint generation to script 7
